Rewrote loadZipDicom to use ImageSeriesReader
Before I was just calling SimpleITK.ReadImage() with a list of dicom
files.  This loads the volume all right, but does not apply the proper
transformation.  And the slices might not be in the proper order.

Calling SimpleITK.ImageSeriesReader.GetGDCMSeriesFileNames()
gets the file names in the proper order, and using the
SimpleITK.ImageSeriesReader class to load the DICOM series
uses the proper transformation.